BUG FIX: No longer using a custom autoloader and moved tests
